Commit 1 - 26/03/2022

+++ routers/routerAPI.js
@@ -1,6 +1,8 @@
 const express = require('express')
 const routerAPI = express.Router()
 
+const utilSec = require('../utils/utilSec')
+
 const knex = require('knex')({
     client: 'pg',
     debug: true,
@@ -10,7 +12,7 @@ const knex = require('knex')({
     }
 });
 
-routerAPI.get('/produtos', (req, res, next) => {
+routerAPI.get('/produtos', utilSec.checkToken, (req, res, next) => {
     knex.select('*').from('produto')
         .then(produtos => res.status(200).json(produtos))
         .catch(err => {
@@ -20,7 +22,7 @@ routerAPI.get('/produtos', (req, res, next) => {
         })
 })
 
-routerAPI.get('/produtos/:id', (req, res, next) => {
+routerAPI.get('/produtos/:id', utilSec.checkToken, (req, res, next) => {
     knex.select('*').from('produto')
         .where({ id: req.params.id })
         .then(produtos => res.status(200).json(produtos))
@@ -31,7 +33,7 @@ routerAPI.get('/produtos/:id', (req, res, next) => {
         })
 })
 
-routerAPI.post('/produtos', (req, res, next) => {
+routerAPI.post('/produtos', utilSec.checkToken, utilSec.isAdmin, (req, res, next) => {
     knex('produto')
         .insert({ descricao: req.body.descricao, valor: req.body.valor, marca: req.body.marca }, ['id'])
         .then(result => {
@@ -43,7 +45,7 @@ routerAPI.post('/produtos', (req, res, next) => {
         })
 })
 
-routerAPI.put('/produtos/:id', (req, res, next) => {
+routerAPI.put('/produtos/:id', utilSec.checkToken, utilSec.isAdmin, (req, res, next) => {
     knex('produto')
         .where({ id: req.params.id })
         .update({ descricao: req.body.descricao, valor: req.body.valor, marca: req.body.marca })
@@ -57,7 +59,7 @@ routerAPI.put('/produtos/:id', (req, res, next) => {
         })
 })
 
-routerAPI.delete('/produtos/:id', (req, res, next) => {
+routerAPI.delete('/produtos/:id', utilSec.checkToken, utilSec.isAdmin, (req, res, next) => {
     knex('produto')
         .where({ id: req.params.id })
         .del()
